refactor: rename table/column-registry to table/registry
Each domain now keeps its registry in a bare registry.ts (core/registry.ts,
effects/registry.ts, table/registry.ts) instead of the table one being the lone
column-registry.ts. Internal path only; exported type names are unchanged.
