+++ resources/js/table/registry.ts
